requirements.txt
flask
requests
peewee
python-dotenv